backend/requirements.txt
flask
flask-cors
psycopg2-binary
pandas
bcrypt
python-dotenv
scikit-learn
xgboost
prophet
joblib
numpy
mlflow
pytest
pytest-flask
requests
prometheus-flask-exporter
prometheus-client
scipy
